refactor: move prints to attach_print
